+++ src/ConsoleTableExporterTest.php
@@ -12,6 +12,8 @@ use Symfony\Component\Console\Output\BufferedOutput;
 use function usleep;
 
 #[CoversClass(ConsoleTableExporter::class)]
+#[CoversClass(TimeUnit::class)]
+#[CoversClass(MemoryUnit::class)]
 final class ConsoleTableExporterTest extends TestCase
 {
     #[Test]