+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `bakame/aide-profiler` will be documented in this file.
 ### Added
 
 - All `Profiler` static methods can have a logger attached to them
+- `MemoryUnit` and `TimeUnit` Enum to ease metrics values conversion.
 
 ### Fixed